+++ make.sh
@@ -2,7 +2,7 @@
 
 function run() {
     echo;
-    LD_LIBRARY_PATH=libPath $1 ./out/main.o;
+    LD_LIBRARY_PATH=$libPath $1 ./out/main.o;
     exit 0;
 }